sanders: Preload GL Drivers
Replacement of Deprecated HWUI Props.
When this property is set to false, Zygote will preload the GL drivers by invoking eglGetDisplay(EGL_DEFAULT_DISPLAY). The goal is to load the dynamic libraries code in Zygote to share it with all the other processes. If a driver does not support being shared, set this property to true.
ref: https://source.android.com/devices/graphics/renderer#properties

Signed-off-by: melvinstark3 <kartik3163@gmail.com>